services/finance-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic[email]==2.9.0
httpx==0.28.1
pytest==8.3.4
pytest-cov==6.0.0
